refactor value assignment out of builder
